refactor: extract page modules (sector.py, company.py) and condense app.py
