+++ assets/custom_styles.css
@@ -30,6 +30,11 @@ h3, h4 {
     color: var(--slate-900);
 }
 
+.btn:hover{
+    background-color: #007bc2;
+    color: white;
+}
+
 .sidebar { 
     background-color: white;
     border-right: 1px solid var(--slate-200);
@@ -99,6 +104,8 @@ h3, h4 {
     flex-direction: column;
     justify-content: center; /* Centers content vertically in large cards */
 }
+
+
 
 /* KPI Value Styling */
 .kpi-value {
